style : usecase 테스트 디렉터리 별로 분리
