+++ app/src/test/java/com/viewpoint/dangder/usecase/auth/LoginUseCaseTest.kt
@@ -1,8 +1,7 @@
-package com.viewpoint.dangder.usecase
+package com.viewpoint.dangder.usecase.auth
 
 import com.google.common.truth.Truth.assertThat
 import com.viewpoint.dangder.repository.AuthRepository
-import com.viewpoint.dangder.usecase.auth.LoginUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.newSingleThreadContext
 import kotlinx.coroutines.test.resetMain
