+++ app/src/test/java/com/viewpoint/dangder/usecase/dog/CheckRegisteredDogUseCaseTest.kt
@@ -1,8 +1,7 @@
-package com.viewpoint.dangder.usecase
+package com.viewpoint.dangder.usecase.dog
 
 import com.google.common.truth.Truth.assertThat
 import com.viewpoint.dangder.repository.DogRepository
-import com.viewpoint.dangder.usecase.dog.CheckRegisteredDogUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.newSingleThreadContext
 import kotlinx.coroutines.test.resetMain
